Backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
openai-whisper
ffmpeg-python
torch
numpy
google-generativeai
python-dotenv
openai